requirements.txt
asgiref==3.7.2
Django==4.2
django-polymorphic==3.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.9.0
